# swagger.yaml
openapi: 3.0.0
info:
  title: Kavoo GO Backend API
  version: 1.0.0
  description: Swagger / OpenAPI spec generated from project routes
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        phone: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        gender: { type: string }
        dateOfBirth: { type: string, format: date }
        country: { type: string }
        city: { type: string }
        employmentStatus: { type: string }
        educationLevel: { type: string }
        maritalStatus: { type: string }
        hasKids: { type: boolean }
        salaryRangeMin: { type: number }
        salaryRangeMax: { type: number }
        hasBankAccount: { type: boolean }
        rewards: { type: number }
        referralCode: { type: string }
...
    Campaign:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        activityType: { type: string }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        minAge: { type: integer }
        maxAge: { type: integer }
        country: { type: string }
        city: { type: string }
        employmentStatus: { type: string }
        educationLevel: { type: string }
        minSalary: { type: number }
        maxSalary: { type: number }
        maritalStatus: { type: string }
        hasKids: { type: boolean }
        rewardAmount: { type: number }
        totalBudget: { type: number }
        costPerUser: { type: number }
        maxUsers: { type: integer }
        videoUrl: { type: string }
        surveyLink: { type: string }
        partnerId: { type: integer }
        status: { type: string }
    Transaction:
      type: object
      properties:
        id: { type: integer }
        transactionId: { type: string }
        date: { type: string, format: date-time }
        status: { type: string }
        campaignId: { type: integer }
        userId: { type: integer }
...
paths:
...
  /api/users/campaigns:
    get:
      tags: [User, Campaign]
      summary: Get campaigns matched to authenticated user (protected - user)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Campaign' }
  /api/users/campaigns/{id}:
    get:
      tags: [User, Campaign]
      summary: Get campaign detail (protected - user)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Campaign detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Campaign' }
...
  /api/partners/campaigns:
    get:
      tags: [Partner, Campaign]
      summary: List campaigns for authenticated partner (protected - partner)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Array of campaigns
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Campaign' }
...
  /api/campaigns/{id}:
    get:
      tags: [Campaign]
      summary: Get campaign by id (protected - dashboard)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Campaign
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Campaign' }
  /api/campaigns:
    post:
      tags: [Campaign]
      summary: Create campaign (protected - dashboard)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                activityType: { type: string }
                startDate: { type: string, format: date-time }
                endDate: { type: string, format: date-time }
                minAge: { type: integer }
                maxAge: { type: integer }
                country: { type: string }
                city: { type: string }
                rewardAmount: { type: number }
                surveyLink: { type: string }
                video: { type: string, format: binary }
      responses:
        '200':
          description: Created campaign
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Campaign' }
...
  /api/admins/users:
    get:
      tags: [Admin, User]
      summary: List all users (protected - dashboard)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Array of users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
  /api/admins/users/{id}:
    get:
      tags: [Admin, User]
      summary: Get user by id including transactions (protected - dashboard)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: User with transactions
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/User'
                  - type: object
                    properties:
                      Transactions:
                        type: array
                        items: { $ref: '#/components/schemas/Transaction' }
  /api/admins/campaigns:
    get:
      tags: [Admin, Campaign]
      summary: List all campaigns (protected - team)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Array of campaigns
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Campaign' }
...